portfolio_cli/requirements.txt
GitPython==3.1.43
PyYAML==6.0.2
markdown==3.6
python-frontmatter==1.1.0
Pillow==10.4.0
python-dateutil==2.9.0.post0

